+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The format is based on [keep a changelog](http://keepachangelog.com) and this pr
 - Ensure that IAP validation for Google rejects pending or cancelled receipts.
 - Fix Lua runtime register_storage_index_filter function.
 - Fix context issue when calling Satori runtime APIs via socket.
+- Fix an issue with message ts precision in returned structs.
 
 ## [3.26.0] - 2025-01-25
 ### Added